remove some warnings

+++ src/rhi/dx12/dx12_pipeline_layout.cpp
@@ -41,8 +41,8 @@ RHI_PIPELINE_LAYOUT* dx12_pipeline_layout_create(const RHI_PIPELINE_LAYOUT_DESC*
 			? &descriptor_ranges_sampler.emplace_back()
 			: &descriptor_ranges.emplace_back();
 		range->RangeType = range_type;
-		range->BaseShaderRegister = descriptor.shader_register_start;
-		range->NumDescriptors = descriptor.shader_register_max;
+		range->BaseShaderRegister = static_cast<UINT>(descriptor.shader_register_start);
+		range->NumDescriptors = static_cast<UINT>(descriptor.shader_register_max);
 		range->RegisterSpace = 0;
 		range->Flags = D3D12_DESCRIPTOR_RANGE_FLAG_DESCRIPTORS_VOLATILE;
 		range->OffsetInDescriptorsFromTableStart = D3D12_DESCRIPTOR_RANGE_OFFSET_APPEND;
